Build with node v22 in CI, bump some dependencies (#7)
**User-Facing Changes**
- Fixes `fzf.Fzf is not a constructor` crash in Raw Messages panel

**Description**
CI was using node v16 to build and upload the final release, and an
incompatibility between the Fzf library and node v16 caused it to embed
the UMD version of the library instead of the ESM version, leading to a
failure at runtime. Building with node v22 should resolve this issue.