Default WebSocket connections to wss:// (#15)
You still need to enter a FQDN for SSL websockets to work, but this
removes the need to add the `wss://` prefix since we assume all robot
websocket connections are via TLS moving forward. To connect to a
non-TLS websocket such as a development server use an explicit `ws://`
prefix.